requirements.txt
# -*- coding: utf-8 -*-
click==8.1.7
kopf==1.37.4
Jinja2==3.1.4
pydantic==2.10.3
PyYAML==6.0.2
kubernetes==31.0.0